Convert scales to H2D tensors in fp8 ops (#1043)
*  Convert scales to H2D tensors in fp8 ops

* Remove invalid header

* Fix other dynamic H2D ops

* Fix lazy float arange in DS

* Fix arange i8, f16 and bf16 lazy cases

* Yet another arange fix

* Review comments

* More review comments

* Create new H2D tensor at each patching to avoid races

* Allocate memory for all H2D scales in one chunk

* Fix gen_op

+++ backend/kernel/hpu_habana_launch_op_pt.cpp
@@ -1602,7 +1602,7 @@ at::Tensor createDynamicTensor(
     synTensorType type) {
   auto allocator = habana::getHABANADeviceAllocator();
   constexpr c10::DispatchKeySet hpu_ks(c10::DispatchKey::HPU);
-  auto dtype = c10::ScalarType::Float;
+  auto dtype = c10::ScalarType::Int;
 
   at::Tensor tensor = at::detail::empty_generic(
       at::asIntArrayRefUnchecked({0}), allocator, hpu_ks, dtype, c10::nullopt);
